refactor: move wsearch utils to wsearch provider
